refactor: move cleanup backend helpers
